requirements.txt
requests
pandas
scikit-learn
sqlalchemy
psycopg2-binary
apache-airflow